# storage/api-docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API AI4BMI",
        "description": "API E-commerce pour l'app mobile (clients). Authentification : Bearer token (Sanctum) ; seuls POST /register et POST /login sont publics.\n\n**Cycle commande** : création (POST /orders) → statut `pending` → paiement confirmé (webhook ou back-office) → `paid` → expédition → `shipped`. Annulation possible uniquement si `pending` (POST /orders/{id}/cancel). Adresse de livraison : fournie à la commande ou reprise du profil (PATCH /user).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "E-commerce"
                ],
                "summary": "Liste des commandes",
                "description": "Commandes de l'utilisateur. Filtres : status (pending, paid, shipped, cancelled), per_page. Réponse paginée.",
                "operationId": "cfbed8ec3108ef0918adcf3d17a8d501",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "paid",
                                "shipped",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OrderResponse"
                                            }
                                        },
                                        "meta": {
                                            "description": "Pagination",
                                            "type": "object"
                                        },
                                        "links": {
                                            "description": "Liens pagination",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "E-commerce"
                ],
                "summary": "Créer une commande",
                "description": "Crée une commande à partir du panier. Adresse de livraison : fournie dans le body ou prise du profil (PATCH /user). Stock vérifié, panier vidé. Statut initial : pending.",
                "operationId": "c9998a283e5013a08ff655c802b61cbd",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "shipping_address": {
                                        "description": "Adresse de livraison (sinon celle du profil)",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "shipping_phone": {
                                        "description": "Téléphone livraison (sinon celui du profil)",
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commande créée (statut pending)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Commande créée."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OrderResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Panier vide ou stock insuffisant"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "E-commerce"
                ],
                "summary": "Détail d'une commande",
                "description": "Détail avec adresse de livraison et lignes (items).",
                "operationId": "b25a937f2f0c4023f41c9e279687033c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OrderResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{id}/cancel": {
            "post": {
                "tags": [
                    "E-commerce"
                ],
                "summary": "Annuler une commande",
                "description": "Annule une commande en statut « pending » uniquement. Le stock des produits est recrédité.",
                "operationId": "bb5b8f90c342274a8ab6969893e46108",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commande annulée (stock recrédité)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OrderResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    },
                    "422": {
                        "description": "Commande non annulable (déjà payée, expédiée ou annulée)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{id}/payment": {
            "post": {
                "tags": [
                    "E-commerce"
                ],
                "summary": "Initier le paiement FedaPay",
                "description": "Pour une commande en statut pending, crée une transaction FedaPay (compte test) et retourne l'URL de paiement. L'app ouvre cette URL (WebView/navigateur) ; après paiement, FedaPay envoie un webhook et la commande passe en paid.",
                "operationId": "4755af94fb8c5f2d7018af9b55a79dcc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "payment_url": {
                                                    "description": "URL à ouvrir pour payer",
                                                    "type": "string"
                                                },
                                                "token": {
                                                    "type": "string"
                                                },
                                                "transaction_id": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    },
                    "422": {
                        "description": "Commande déjà payée ou non pending / FedaPay non configuré"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "OrderResponse": {
                "description": "Commande avec adresse de livraison et lignes",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 99.5
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "paid",
                            "shipped",
                            "cancelled"
                        ]
                    },
                    "shipping_address": {
                        "type": "string",
                        "nullable": true
                    },
                    "shipping_phone": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "items": {
                        "description": "Lignes de la commande (product, quantity, price, subtotal)",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}